Move some traits from exchanges::traits
