+++ backend/libs/exchanges/Cargo.toml
@@ -9,8 +9,10 @@ edition = "2018"
 [dependencies]
 rpc = { path = "../rpc" }
 config = { path = "../config" }
-async-trait = "0.1.42"
 types = { path = "../types" }
+base_recorder = { path = "../rulesets/base_recorder" }
+
+async-trait = "0.1.42"
 reqwest = {version = "0.11.0", features = ["rustls-tls-native-roots", "json", "gzip", "brotli"]}
 tokio = { version = "1.2", features = [
   "rt", "rt-multi-thread", "sync", "macros", "time", "net"
