+++ backend/libs/exchanges/src/traits.rs
@@ -23,79 +23,6 @@ use super::entities::{
 use super::errors::ExecutionFailed;
 
 #[async_trait]
-pub trait Recorder {
-  fn get_database(&self) -> &Database;
-  fn get_col_name(&self) -> &str;
-  async fn update_indices(&self, flds: &[&str]) {
-    let col_name = self.get_col_name();
-    let db = self.get_database();
-    let has_index = db
-      .run_command(doc! {"listIndexes": &col_name}, None)
-      .await
-      .map(|item| {
-        return item
-          .get_document("listIndexes")
-          .unwrap_or(&doc! {"ok": false})
-          .get_bool("ok")
-          .unwrap_or(false);
-      })
-      .unwrap_or(false);
-    if !has_index {
-      let mut indices = vec![];
-      for fld_name in flds {
-        indices.push(doc! { "name": format!("{}_index", fld_name), "key": doc!{
-          *fld_name: 1,
-        } })
-      }
-      let _ = db
-        .run_command(
-          doc! {
-            "createIndexes": &col_name,
-            "indexes": indices
-          },
-          None,
-        )
-        .await;
-    }
-  }
-}
-
-#[async_trait]
-pub trait HistoryFetcher {
-  async fn refresh(
-    &self,
-    symbols: Vec<String>,
-  ) -> ThreadSafeResult<Subscription>;
-  async fn stop(&self) -> ThreadSafeResult<()>;
-  async fn spawn(&self) -> ThreadSafeResult<()>;
-}
-
-#[async_trait]
-pub trait SymbolFetcher {
-  async fn refresh(&self) -> ThreadSafeResult<()>;
-}
-
-#[async_trait]
-pub trait HistoryRecorder {
-  async fn spawn(&self);
-}
-
-#[async_trait]
-pub trait SymbolRecorder {
-  type ListStream: Stream + Send + 'static;
-  async fn list(
-    &self,
-    query: impl Into<Option<Document>> + Send + 'async_trait,
-  ) -> ThreadSafeResult<Self::ListStream>;
-  async fn update_symbols<T>(
-    &self,
-    value: Vec<T>,
-  ) -> ThreadSafeResult<InsertManyResult>
-  where
-    T: Serialize + Send;
-}
-
-#[async_trait]
 pub trait TradeObserver {
   async fn start(&self) -> GenericResult<()>;
   async fn subscribe(&self) -> ::std::io::Result<BoxStream<'_, BookTicker>>;
