+++ backend/libs/rulesets/base_recorder/Cargo.toml
@@ -7,3 +7,6 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+async-trait = "0.1.42"
+futures = "0.3.12"
+mongodb = {version = "1.1.1", default-features = false, features = ["async-std-runtime"]}
