+++ backend/libs/rulesets/history_fetcher/Cargo.toml
@@ -7,3 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+types = { path = "../../types" }
+nats = "0.8"
+
+async-trait = "0.1.42"
