+++ backend/libs/rulesets/history_fetcher/src/lib.rs
@@ -1,7 +1,14 @@
-#[cfg(test)]
-mod tests {
-  #[test]
-  fn it_works() {
-    assert_eq!(2 + 2, 4);
-  }
+use ::async_trait::async_trait;
+use ::nats::asynk::Subscription;
+
+use ::types::ThreadSafeResult;
+
+#[async_trait]
+pub trait HistoryFetcher {
+  async fn refresh(
+    &self,
+    symbols: Vec<String>,
+  ) -> ThreadSafeResult<Subscription>;
+  async fn stop(&self) -> ThreadSafeResult<()>;
+  async fn spawn(&self) -> ThreadSafeResult<()>;
 }
