+++ backend/libs/rulesets/history_recorder/src/lib.rs
@@ -1,7 +1,6 @@
-#[cfg(test)]
-mod tests {
-  #[test]
-  fn it_works() {
-    assert_eq!(2 + 2, 4);
-  }
+use ::async_trait::async_trait;
+
+#[async_trait]
+pub trait HistoryRecorder {
+  async fn spawn(&self);
 }
