+++ backend/libs/rulesets/symbol_fetcher/Cargo.toml
@@ -7,3 +7,6 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+types = { path = "../../types" }
+
+async-trait = "0.1.42"
