+++ backend/libs/rulesets/symbol_fetcher/src/lib.rs
@@ -1,7 +1,8 @@
-#[cfg(test)]
-mod tests {
-  #[test]
-  fn it_works() {
-    assert_eq!(2 + 2, 4);
-  }
+use ::async_trait::async_trait;
+
+use ::types::ThreadSafeResult;
+
+#[async_trait]
+pub trait SymbolFetcher {
+  async fn refresh(&self) -> ThreadSafeResult<()>;
 }
