+++ backend/libs/rulesets/symbol_recorder/src/lib.rs
@@ -1,7 +1,22 @@
-#[cfg(test)]
-mod tests {
-  #[test]
-  fn it_works() {
-    assert_eq!(2 + 2, 4);
-  }
+use ::async_trait::async_trait;
+use ::futures_core::Stream;
+use ::mongodb::bson::Document;
+use ::mongodb::results::InsertManyResult;
+use ::serde::Serialize;
+
+use ::types::ThreadSafeResult;
+
+#[async_trait]
+pub trait SymbolRecorder {
+  type ListStream: Stream + Send + 'static;
+  async fn list(
+    &self,
+    query: impl Into<Option<Document>> + Send + 'async_trait,
+  ) -> ThreadSafeResult<Self::ListStream>;
+  async fn update_symbols<T>(
+    &self,
+    value: Vec<T>,
+  ) -> ThreadSafeResult<InsertManyResult>
+  where
+    T: Serialize + Send;
 }
