+++ backend/libs/types/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2018"
 
 [dependencies]
 url = "2.2.0"
-serde = "1.0.123"
+serde = {version = "1.0.123", features = ["derive"]}
 tokio = {version = "1.2", features = ["sync"]}
 chrono = "0.4.19"
 warp = "0.3"
